backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# API client
requests>=2.28.0

# CORS support
starlette>=0.27.0

# For production deployment (optional)
# gunicorn>=21.0.0
# uvloop>=0.17.0
# httptools>=0.5.0